Add a transaction pooling "queue" for IDB storer
Significantly improves throughput.
Tradeoff is now any error can cause failures
for multiple other independent file operations.

This could be problematic, but as long as the
chance of errors inside the storer are reduced,
this should not be an issue. Currently only
error is if the parent directory does not exist.

This created a speedup of first-boot time from
45 seconds down to 22 seconds!